Refactoring, resolve level 6
