+++ phpstan.neon
@@ -7,7 +7,7 @@ parameters:
     - src
 
   # Level 9 is max
-  level: 5
+  level: 6
 
   ignoreErrors:
     - '#Call to an undefined method ReflectionType::isBuiltin\(\)\.#'
