+++ src/Contracts/Renderer.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Soyhuce\NextIdeHelper\Contracts;
+
+interface Renderer
+{
+    public function render(): void;
+}
