+++ src/Domain/Models/Actions/ResolveModelScopes.php
@@ -5,6 +5,7 @@ namespace Soyhuce\NextIdeHelper\Domain\Models\Actions;
 use Illuminate\Support\Str;
 use ReflectionClass;
 use ReflectionMethod;
+use Soyhuce\NextIdeHelper\Contracts\ModelResolver;
 use Soyhuce\NextIdeHelper\Domain\Models\Entities\Model;
 use Soyhuce\NextIdeHelper\Entities\Method;
 use Soyhuce\NextIdeHelper\Support\Reflection\FunctionReflection;
