+++ src/Domain/Models/Entities/QueryBuilder.php
@@ -21,7 +21,7 @@ class QueryBuilder
         $this->extras = collect();
     }
 
-    public function isBuiltIn()
+    public function isBuiltIn(): bool
     {
         return $this->fqcn === '\\' . EloquentBuilder::class;
     }
