+++ src/Domain/Models/Output/ModelDocBlock.php
@@ -5,13 +5,14 @@ namespace Soyhuce\NextIdeHelper\Domain\Models\Output;
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
+use Soyhuce\NextIdeHelper\Contracts\Renderer;
 use Soyhuce\NextIdeHelper\Domain\Models\Entities\Attribute;
 use Soyhuce\NextIdeHelper\Domain\Models\Entities\Model;
 use Soyhuce\NextIdeHelper\Domain\Models\Entities\Relation;
 use Soyhuce\NextIdeHelper\Support\Output\DocBlock;
 use function in_array;
 
-class ModelDocBlock extends DocBlock
+class ModelDocBlock extends DocBlock implements Renderer
 {
     private Model $model;
